+++ test-integration/servers-error-log.test.js
@@ -7,7 +7,7 @@ const path = require('path');
 const {expect} = require('chai');
 
 const Server = require('../test/servers/server');
-const {makeTestGenerator} = require('../test/helpers');
+const {makeTestGenerator} = require('./_helpers');
 
 const {
   Checkers,
