+++ test-integration/servers-log.test.js
@@ -7,7 +7,7 @@ const {expect} = require('chai');
 const semver = require('semver');
 
 const Server = require('../test/servers/server');
-const {makeTestGenerator} = require('../test/helpers');
+const {makeTestGenerator} = require('./_helpers');
 
 const {
   Checkers,
